Manage adding additional maneuvers to existing counts
In the original, if we have 1 light and 1 wind, setting `wind 2` would mean we had 1 light 1 wind 2 wind i.e. 4 total maneuvers, which is invalid.

This is not logical behaviour to the user.

This change disregards the existing value for any new maneuver and counts only the new value for validation of totals.

As long as the total is not exceeded the new value replaces the old value.

It also sets n to zero so we revert to using the original commandArgs value