Remove get.maneuver and get.ids
Another holdover from Singer?

I think this used to look up the JA id based on the long name and use that to check recast. But all maneuvers share a recast so this is pointless.

We can just get the long name (to send to the client) from get.maneuvers