Register events for gaining/losing buffs
Was fed-up of exiting a HTMB and going AFK then finding, 2 hours later, I was stood there doing maneuvers. Not in anyway suspicious!

The issues is that when you exit some battlefields you don't actually zone. So, this looks at the buff you lose when you "exit" the battle. Works for in-zone content like Voidwatch, Reive and Campaign. Hope to make this more flexible.

Also, having been lucky not to die in Odyssey, AutoPUP will now stop when you gain the "Invisible" buff.